Battlemap zones render properly now; added 'gappy'
Gappy means the floor will be rendered as though any 'm' are meant to be
gaps in a border, but the floor should be solid.  Basically, gappy means
take all but the first m and make it a line.  Usually the default of 'true'
will be enough, but occasionaly gappiness is not desired, like a bunch of
boxes.